[hvac] Fix Temp Up appearing to do nothing while Temp Down works [issue tracker] N/A [description] Change HvacListener.onTempChanged() from int to float end-to-end -- the interface, HvacManager's dispatch site, and HvacViewModel's implementation. [symptom] Temp Down looked correct every time; Temp Up looked like it did nothing. incrementTemp()/decreaseTemp() step in 0.5C, so every increment lands on X.5. [rootcause] HvacListener.onTempChanged(int value, int area) took an int, so HvacManager's dispatcher truncated the echoed float toward zero ((int) event.getValue()) before it ever reached the ViewModel. Truncating X.5 toward zero always rounds the confirmed value back down to X, silently cancelling the +0.5 the instant the echo arrived, while decrements (landing on X.5, then truncating to a lower whole number each time) always showed a net decrease -- matching the observed Up-looks-broken/Down-looks-fine asymmetry. [action] Changed onTempChanged to take a float in HvacListener.java, HvacManager.java's dispatch site (dropped the (int) cast), and HvacViewModel.java's implementation, so the real 0.5-degree value survives the round trip instead of being cast to int along the way. [test step] Manual: tapped Temp Up repeatedly via the app and watched the confirmed (echoed) value in logcat/UI track the optimistic 0.5-degree steps instead of rounding back down. [test rate] pass

+++ kpit/automotive/hmi/hvac_app/src/com/kpit/hmi/hvac/viewmodel/HvacViewModel.java
@@ -162,7 +162,7 @@ public class HvacViewModel extends AndroidViewModel implements HvacListener, Sys
     }
 
     @Override
-    public void onTempChanged(int value, int area) {
+    public void onTempChanged(float value, int area) {
         if (!mCachedOtherControlsEnable) return;
         if (area == 1) {
             mLeftTemp = value;
